pyproject.toml
[project]
name = "grid_trading_bot"
version = "0.1.0"
description = "Open-source cryptocurrency trading bot designed to perform grid trading strategies using historical data for backtesting."
authors = [{ name = "Jordan TETE", email = "tetej171@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE.txt" }
requires-python = ">=3.12"
keywords = ["grid-trading", "trading-bot", "cryptocurrency", "backtesting", "ccxt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pandas==2.2.3",
    "numpy==2.1.3",
    "plotly==6.5.2",
    "tabulate==0.9.0",
    "aiohttp==3.13.3",
    "apprise==1.9.7",
    "ccxt==4.5.37",
    "configparser==7.2.0",
    "psutil==7.2.2",
    "python-dotenv==1.2.1",
    "click>=8.1"
]

[project.scripts]
grid_trading_bot = "grid_trading_bot.cli:main"

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "pytest-timeout==2.4.0",
    "pre-commit==4.5.1",
]

[project.urls]
repository= "https://github.com/jordantete/grid_trading_bot"
issues = "https://github.com/jordantete/grid_trading_bot/issues"
discussions = "https://github.com/jordantete/grid_trading_bot/discussions"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
log_cli = true
log_cli_level = "INFO"
addopts = "-m 'not sandbox'"
markers = [
    "asyncio: mark a test as an async test",
    "timeout: mark a test with a timeout",
    "integration: end-to-end integration tests (backtest)",
    "sandbox: tests requiring network access to exchange (skipped by default)",
]

[tool.coverage.run]
source = ["grid_trading_bot"]
omit = [
    "*/interface*.py"
]

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src/grid_trading_bot", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "SIM",  # flake8-simplify
    "RUF",  # Ruff-specific rules
    "PT",   # flake8-pytest-style
    "Q",    # flake8-quotes
    "A",    # flake8-builtins
    "COM",  # flake8-commas
    "DTZ",  # flake8-datetimez
    "TCH",  # flake8-type-checking
]
ignore = [
    "S101",   # assert used (OK in tests)
    "COM812", # missing trailing comma (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint.isort]
known-first-party = ["grid_trading_bot"]
force-sort-within-sections = true

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true
